refactor(zfs): restructure code
